PAPP-9543 support concurrent deflate item actions
Several concurrent runs of "deflate item" were causing action failures
due to the tmp file created in add_file_to_vault() being deleted by
one action run while another was still trying to use it. This fix is
to append a 16 character random string to the end of the file path in
order to give each action run a unique file to work with.